Support large harddisks. 4TB now works.

+++ lintory/models.py
@@ -481,7 +481,7 @@ class hardware_type(generic_type):
 
 class storage(hardware):
     used_by  = models.ForeignKey('computer',related_name='used_storage',null=True,blank=True)
-    total_size = models.PositiveIntegerField(null=True,blank=True)
+    total_size = models.DecimalField(max_digits=12,decimal_places=0,null=True,blank=True)
     sector_size = models.PositiveIntegerField(null=True,blank=True)
 
     def os_list(self):
